src/main/java/mathandel/backend/client/request/ProductDataRequest.java
package mathandel.backend.client.request;

import org.apache.commons.lang3.builder.EqualsBuilder;
import org.apache.commons.lang3.builder.HashCodeBuilder;

public class ProductDataRequest {

    private String name;
    private String description;

    public ProductDataRequest() {
    }

    public String getName() {
        return name;
    }

    public ProductDataRequest setName(String name) {
        this.name = name;
        return this;
    }

    public String getDescription() {
        return description;
    }

    public ProductDataRequest setDescription(String description) {
        this.description = description;
        return this;
    }

    @Override
    public boolean equals(Object o) {
        if (this == o) {
            return true;
        }

        if (o == null || getClass() != o.getClass()) {
            return false;
        }

        ProductDataRequest that = (ProductDataRequest) o;

        return new EqualsBuilder()
                .append(name, that.name)
                .append(description, that.description)
                .isEquals();
    }

    @Override
    public int hashCode() {
        return new HashCodeBuilder()
                .append(name)
                .append(description)
                .toHashCode();
    }
}
